fixes after bumping version
